Pull variable creation logic out of `ops/variable_scope.py`.
Since `variable_scope.py` is currently responsible for both variable scoping and much of the variable creation logic, it has dependencies on `variable_v1.py` and `ref_variable.py` that are causing cycles. This change involves pulling the variable creation logic and proto registrations out of `variable_scope.py` and into `ref_variable.py` and `resource_variable_ops.py` (for tf1 and tf2 respectively). This is part one of two changes to remove `variable_scope`'s explicit dependency on variable_v1 logic, which breaks many import cycles.

A lazy load in `ref_variable.py` is needed while the refactors are in progress.

PiperOrigin-RevId: 527488606
