+++ tensorflow/python/BUILD
@@ -172,6 +172,7 @@ py_library(
         ":proto_ops",
         ":pywrap_tensorflow",
         ":pywrap_tfe",
+        ":ref_variable",
         ":rnn_ops_gen",
         ":script_ops",
         ":sendrecv_ops_gen",
@@ -184,6 +185,7 @@ py_library(
         ":tensor_array_ops",
         ":training",
         ":uniform_quant_ops_gen",
+        ":variable_v1",
         ":weights_broadcast_ops",
         ":while_v2",
         "//tensorflow/core:protos_all_py",
@@ -3255,6 +3257,7 @@ py_library(
     deps = [
         ":array_ops",
         ":array_ops_gen",
+        ":resource_variable_ops",
         ":state_ops",
         ":state_ops_gen",
         ":variable_v1",
@@ -3270,6 +3273,7 @@ py_library(
         "//tensorflow/python/types",
         "//tensorflow/python/util:compat",
         "//tensorflow/python/util:deprecation",
+        "//tensorflow/python/util:lazy_loader",
     ],
 )
